pyproject.toml
[tool.poetry]
name = "drive-organizer-api"
version = "1.0.0"
description = "Google Drive Organizer API"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "apps/api"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
httpx = "^0.25.2"
google-api-python-client = "^2.108.0"
google-auth = "^2.23.4"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
openai = "^1.3.7"
supabase = "^2.0.2"
structlog = "^23.2.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
ruff = "^0.1.6"
black = "^23.11.0"
isort = "^5.12.0"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.black]
target-version = ['py312']
line-length = 88

[tool.isort]
profile = "black"
multi_line_output = 3 